AA-1324-payment-summary-plastic-card
- Adds SummeryFooterView
- Refactors package hierarchy to have all the Summary blocks in the same
package.
