+++ library/src/main/java/com/spendesk/grapes/component/content/summary/block/SummaryBlockContentInlineView.kt
@@ -1,20 +1,20 @@
-package com.spendesk.grapes.component.content.summary.noneditable
+package com.spendesk.grapes.component.content.summary.block
 
 import android.content.Context
 import android.util.AttributeSet
 import android.view.View
 import com.spendesk.grapes.R
-import com.spendesk.grapes.component.content.summary.SummaryBlockTitleView
-import com.spendesk.grapes.component.content.summary.SummaryBlockView
+import com.spendesk.grapes.component.content.summary.block.definition.SummaryBlockTitleView
+import com.spendesk.grapes.component.content.summary.block.definition.SummaryBlockView
 import com.spendesk.grapes.list.content.summary.SummaryBlockContentAdapter
 import com.spendesk.grapes.list.content.summary.SummaryBlockContentModel
-import kotlinx.android.synthetic.main.summary_block_content_approver.view.*
+import kotlinx.android.synthetic.main.summary_block_content_inline.view.*
 
 /**
  * @author Vivien Mahe
  * @since 07/01/2021
  */
-class SummaryBlockContentApproverView : SummaryBlockView {
+class SummaryBlockContentInlineView : SummaryBlockView {
 
     //region constructors
     constructor(context: Context) : super(context)
@@ -24,18 +24,18 @@ class SummaryBlockContentApproverView : SummaryBlockView {
 
     class Configuration(
         titleConfiguration: SummaryBlockTitleView.Configuration,
-        val items: List<SummaryBlockContentModel.ApproverStatus>
+        val items: List<SummaryBlockContentModel.InlineKeyValue>
     ) : SummaryBlockView.Configuration(titleConfiguration)
 
     private val adapter = SummaryBlockContentAdapter()
 
     init {
-        View.inflate(context, R.layout.summary_block_content_approver, this)
+        View.inflate(context, R.layout.summary_block_content_inline, this)
 
-        summaryBlockContentApproverList.adapter = adapter
+        summaryBlockContentInlineList.adapter = adapter
     }
 
-    override fun getSummaryBlockTitleView(): SummaryBlockTitleView = summaryBlockContentApproverTitle
+    override fun getSummaryBlockTitleView(): SummaryBlockTitleView = summaryBlockContentInlineTitle
 
     fun updateConfiguration(configuration: Configuration) {
         super.updateConfiguration(configuration)
