+++ library/src/main/java/com/spendesk/grapes/component/content/summary/block/SummaryBlockContentMapView.kt
@@ -1,12 +1,12 @@
-package com.spendesk.grapes.component.content.summary.noneditable
+package com.spendesk.grapes.component.content.summary.block
 
 import android.content.Context
 import android.util.AttributeSet
 import android.view.View
 import androidx.core.view.isVisible
 import com.spendesk.grapes.R
-import com.spendesk.grapes.component.content.summary.SummaryBlockTitleView
-import com.spendesk.grapes.component.content.summary.SummaryBlockView
+import com.spendesk.grapes.component.content.summary.block.definition.SummaryBlockTitleView
+import com.spendesk.grapes.component.content.summary.block.definition.SummaryBlockView
 import com.spendesk.grapes.extensions.afterMeasured
 import com.spendesk.grapes.extensions.gone
 import com.spendesk.grapes.extensions.visible
