+++ library/src/main/java/com/spendesk/grapes/component/content/summary/block/SummaryBlockContentResponsibilityCenterGaugeLegendView.kt
@@ -1,4 +1,4 @@
-package com.spendesk.grapes.component.content.summary.noneditable
+package com.spendesk.grapes.component.content.summary.block
 
 import android.content.Context
 import android.util.AttributeSet
