+++ library/src/main/res/drawable-anydpi/ic_payment_type_plastic_card.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="16dp"
+    android:height="16dp"
+    android:viewportWidth="16"
+    android:viewportHeight="16">
+  <path
+      android:pathData="M1.3333,2.6667H14.6667C15.4026,2.6677 15.999,3.2641 16,4.0001V4.889C16,5.1344 15.801,5.3334 15.5556,5.3334H0.4444C0.199,5.3334 0,5.1344 0,4.889V4.0001C0.001,3.2641 0.5974,2.6677 1.3333,2.6667ZM16,12.0001V7.5556C16,7.3102 15.801,7.1112 15.5556,7.1112H0.4444C0.199,7.1112 0,7.3102 0,7.5556V12.0001C0.001,12.7361 0.5974,13.3324 1.3333,13.3334H14.6667C15.4026,13.3324 15.999,12.7361 16,12.0001Z"
+      android:fillColor="#706F81"
+      android:fillType="evenOdd"/>
+</vector>
