+++ library/src/main/res/drawable-anydpi/ic_request_type_card_load.xml
