+++ library/src/main/res/drawable-anydpi/ic_request_type_subscription.xml
