+++ library/src/main/res/layout/summary_footer.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="utf-8"?>
+<merge xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:parentTag="androidx.constraintlayout.widget.ConstraintLayout">
+
+    <TextView
+        android:id="@+id/summaryFooterTitle"
+        style="@style/SummaryFooterTitleText"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="@dimen/summaryFooterMarginHorz"
+        android:layout_marginTop="@dimen/summaryFooterTitleMarginTop"
+        android:layout_marginEnd="@dimen/summaryFooterMarginHorz"
+        app:layout_constraintBottom_toTopOf="@id/summaryFooterSubtitle"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        tools:text="Transaction description" />
+
+    <TextView
+        android:id="@+id/summaryFooterSubtitle"
+        style="@style/SummaryFooterSubtitleText"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="@dimen/summaryFooterMarginHorz"
+        android:layout_marginTop="@dimen/summaryFooterSubtitleMarginTop"
+        android:layout_marginEnd="@dimen/summaryFooterMarginHorz"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/summaryFooterTitle"
+        tools:text="Mich Much -278240952REF 33 Sir John Rogerson's Quay Upwork.com/bi Dublin 2 IRL" />
+</merge>
