+++ library/src/main/res/layout/summary_header.xml
@@ -6,11 +6,11 @@
 
     <com.spendesk.grapes.UserSupplierInlineView
         android:id="@+id/summaryHeaderExpenseAmountUserSupplier"
-        style="@style/TemplateHeaderExpenseUserSupplier"
+        style="@style/SummaryHeaderExpenseUserSupplier"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/templateHeaderExpenseMarginHorz"
-        android:layout_marginEnd="@dimen/templateHeaderExpenseMarginHorz"
+        android:layout_marginStart="@dimen/summaryHeaderExpenseMarginHorz"
+        android:layout_marginEnd="@dimen/summaryHeaderExpenseMarginHorz"
         app:layout_constraintBottom_toTopOf="@id/summaryHeaderExpenseAmountTitle"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -22,12 +22,12 @@
 
     <TextView
         android:id="@+id/summaryHeaderExpenseAmountTitle"
-        style="@style/TemplateHeaderExpenseAmountTitleText"
+        style="@style/SummaryHeaderExpenseAmountTitleText"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/templateHeaderExpenseMarginHorz"
-        android:layout_marginTop="@dimen/templateHeaderExpenseAmountTitleMarginTop"
-        android:layout_marginEnd="@dimen/templateHeaderExpenseMarginHorz"
+        android:layout_marginStart="@dimen/summaryHeaderExpenseMarginHorz"
+        android:layout_marginTop="@dimen/summaryHeaderExpenseAmountTitleMarginTop"
+        android:layout_marginEnd="@dimen/summaryHeaderExpenseMarginHorz"
         app:layout_constraintBottom_toTopOf="@id/summaryHeaderExpenseAmountSubtitle"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -36,12 +36,12 @@
 
     <TextView
         android:id="@+id/summaryHeaderExpenseAmountSubtitle"
-        style="@style/TemplateHeaderExpenseAmountSubtitleText"
+        style="@style/SummaryHeaderExpenseAmountSubtitleText"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/templateHeaderExpenseMarginHorz"
-        android:layout_marginTop="@dimen/templateHeaderExpenseAmountSubtitleMarginTop"
-        android:layout_marginEnd="@dimen/templateHeaderExpenseMarginHorz"
+        android:layout_marginStart="@dimen/summaryHeaderExpenseMarginHorz"
+        android:layout_marginTop="@dimen/summaryHeaderExpenseAmountSubtitleMarginTop"
+        android:layout_marginEnd="@dimen/summaryHeaderExpenseMarginHorz"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/summaryHeaderExpenseAmountTitle"
@@ -49,12 +49,12 @@
 
     <TextView
         android:id="@+id/summaryHeaderExpenseDescriptionTitle"
-        style="@style/TemplateHeaderExpenseDescriptionTitleText"
+        style="@style/SummaryHeaderExpenseDescriptionTitleText"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/templateHeaderExpenseMarginHorz"
-        android:layout_marginTop="@dimen/templateHeaderExpenseDescriptionMarginTop"
-        android:layout_marginEnd="@dimen/templateHeaderExpenseMarginHorz"
+        android:layout_marginStart="@dimen/summaryHeaderExpenseMarginHorz"
+        android:layout_marginTop="@dimen/summaryHeaderExpenseDescriptionMarginTop"
+        android:layout_marginEnd="@dimen/summaryHeaderExpenseMarginHorz"
         app:layout_constraintBottom_toTopOf="@id/summaryHeaderExpenseDescriptionSubtitle"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -63,12 +63,12 @@
 
     <TextView
         android:id="@+id/summaryHeaderExpenseDescriptionSubtitle"
-        style="@style/TemplateHeaderExpenseDescriptionSubtitleText"
+        style="@style/SummaryHeaderExpenseDescriptionSubtitleText"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/templateHeaderExpenseMarginHorz"
-        android:layout_marginTop="@dimen/templateHeaderExpenseDescriptionSubtitleMarginTop"
-        android:layout_marginEnd="@dimen/templateHeaderExpenseMarginHorz"
+        android:layout_marginStart="@dimen/summaryHeaderExpenseMarginHorz"
+        android:layout_marginTop="@dimen/summaryHeaderExpenseDescriptionSubtitleMarginTop"
+        android:layout_marginEnd="@dimen/summaryHeaderExpenseMarginHorz"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/summaryHeaderExpenseDescriptionTitle"
@@ -78,9 +78,9 @@
         android:id="@+id/summaryHeaderExpenseTypeMessage"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/templateHeaderExpenseMarginHorz"
-        android:layout_marginTop="@dimen/templateHeaderExpenseTypeMessageMarginTop"
-        android:layout_marginEnd="@dimen/templateHeaderExpenseMarginHorz"
+        android:layout_marginStart="@dimen/summaryHeaderExpenseMarginHorz"
+        android:layout_marginTop="@dimen/summaryHeaderExpenseTypeMessageMarginTop"
+        android:layout_marginEnd="@dimen/summaryHeaderExpenseMarginHorz"
         app:layout_constraintBottom_toTopOf="@id/summaryHeaderExpenseDenyReasonBlock"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/summaryHeaderExpenseDescriptionSubtitle"
@@ -90,12 +90,12 @@
 
     <com.spendesk.grapes.messages.MessageBlockView
         android:id="@+id/summaryHeaderExpenseDenyReasonBlock"
-        style="@style/TemplateHeaderExpenseDenyReasonMessage"
+        style="@style/SummaryHeaderExpenseDenyReasonMessage"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/templateHeaderExpenseMarginHorz"
-        android:layout_marginTop="@dimen/templateHeaderExpenseDenyReasonMessageMarginTop"
-        android:layout_marginEnd="@dimen/templateHeaderExpenseMarginHorz"
+        android:layout_marginStart="@dimen/summaryHeaderExpenseMarginHorz"
+        android:layout_marginTop="@dimen/summaryHeaderExpenseDenyReasonMessageMarginTop"
+        android:layout_marginEnd="@dimen/summaryHeaderExpenseMarginHorz"
         app:layout_constraintBottom_toTopOf="@id/summaryHeaderExpenseBottomSeparator"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -105,10 +105,10 @@
 
     <View
         android:id="@+id/summaryHeaderExpenseBottomSeparator"
-        style="@style/TemplateHeaderExpenseBottomSeparator"
+        style="@style/SummaryHeaderExpenseBottomSeparator"
         android:layout_width="0dp"
-        android:layout_height="@dimen/templateHeaderExpenseBottomSeparatorSize"
-        android:layout_marginTop="@dimen/templateHeaderExpenseBottomSeparatorMarginTop"
+        android:layout_height="@dimen/summaryHeaderExpenseBottomSeparatorSize"
+        android:layout_marginTop="@dimen/summaryHeaderExpenseBottomSeparatorMarginTop"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/summaryHeaderExpenseDenyReasonBlock" />
