--- library/src/main/res/values/dimens_template_header.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <!-- region Header Expense -->
-    <dimen name="templateHeaderExpenseMarginHorz">16dp</dimen>
-
-    <dimen name="templateHeaderExpenseAmountTitleMarginTop">28dp</dimen>
-    <dimen name="templateHeaderExpenseAmountTitleTextSize">40sp</dimen>
-    <dimen name="templateHeaderExpenseAmountSubtitleMarginTop">4dp</dimen>
-    <dimen name="templateHeaderExpenseAmountSubtitleTextSize">16sp</dimen>
-
-    <dimen name="templateHeaderExpenseDescriptionMarginTop">24dp</dimen>
-    <dimen name="templateHeaderExpenseDescriptionTitleTextSize">24sp</dimen>
-    <dimen name="templateHeaderExpenseDescriptionSubtitleMarginTop">8dp</dimen>
-    <dimen name="templateHeaderExpenseDescriptionSubtitleTextSize">16sp</dimen>
-
-    <dimen name="templateHeaderExpenseTypeMessageMarginTop">40dp</dimen>
-
-    <dimen name="templateHeaderExpenseDenyReasonMessageMarginTop">@dimen/normalMargin</dimen>
-
-    <dimen name="templateHeaderExpenseBottomSeparatorMarginTop">16dp</dimen>
-    <dimen name="templateHeaderExpenseBottomSeparatorSize">0.5dp</dimen>
-    <!-- endregion Header Expense -->
-</resources>
