+++ library/src/main/res/values/styles_summary_footer.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <style name="SummaryFooterTitleText">
+        <item name="android:textColor">@color/summaryFooterTitleText</item>
+        <item name="android:textSize">@dimen/summaryFooterTitleTextSize</item>
+        <item name="android:fontFamily">@font/gt_america_bold</item>
+        <item name="android:singleLine">true</item>
+        <item name="android:gravity">center_horizontal</item>
+    </style>
+
+    <style name="SummaryFooterSubtitleText">
+        <item name="android:textColor">@color/summaryFooterSubtitleText</item>
+        <item name="android:textSize">@dimen/summaryFooterSubtitleTextSize</item>
+        <item name="android:fontFamily">@font/gt_america</item>
+        <item name="android:gravity">center_horizontal</item>
+    </style>
+</resources>
