--- library/src/main/res/values/styles_template_header.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <!-- region Header Expense -->
-    <style name="TemplateHeaderExpenseUserSupplier">
-        <item name="android:visibility">gone</item>
-    </style>
-
-    <style name="TemplateHeaderExpenseAmountTitleText">
-        <item name="android:textColor">@color/templateHeaderExpenseAmountTitleText</item>
-        <item name="android:textSize">@dimen/templateHeaderExpenseAmountTitleTextSize</item>
-        <item name="android:fontFamily">@font/gt_america</item>
-        <item name="android:singleLine">true</item>
-    </style>
-
-    <style name="TemplateHeaderExpenseAmountSubtitleText">
-        <item name="android:textColor">@color/templateHeaderExpenseAmountSubtitleText</item>
-        <item name="android:textSize">@dimen/templateHeaderExpenseAmountSubtitleTextSize</item>
-        <item name="android:fontFamily">@font/gt_america</item>
-        <item name="android:singleLine">true</item>
-    </style>
-
-    <style name="TemplateHeaderExpenseDescriptionTitleText">
-        <item name="android:textColor">@color/templateHeaderExpenseDescriptionTitleText</item>
-        <item name="android:textSize">@dimen/templateHeaderExpenseDescriptionTitleTextSize</item>
-        <item name="android:fontFamily">@font/gt_america</item>
-    </style>
-
-    <style name="TemplateHeaderExpenseDescriptionSubtitleText">
-        <item name="android:textColor">@color/templateHeaderExpenseDescriptionSubtitleText</item>
-        <item name="android:textSize">@dimen/templateHeaderExpenseDescriptionSubtitleTextSize</item>
-        <item name="android:fontFamily">@font/gt_america</item>
-    </style>
-
-    <style name="TemplateHeaderExpenseDenyReasonMessage">
-        <item name="messageStyle">alert</item>
-        <item name="android:visibility">gone</item>
-    </style>
-
-    <style name="TemplateHeaderExpenseBottomSeparator">
-        <item name="android:background">@color/templateHeaderExpenseBottomSeparator</item>
-    </style>
-
-    <!-- endregion Header Expense -->
-
-</resources>
