+++ sample/src/main/res/layout/fragment_home_contents.xml
@@ -53,7 +53,7 @@
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toTopOf="parent" />
 
-                <com.spendesk.grapes.component.content.summary.noneditable.SummaryBlockContentMapView
+                <com.spendesk.grapes.component.content.summary.block.SummaryBlockContentMapView
                     android:id="@+id/homeButtonSectionMapExtendedBlock"
                     android:layout_width="0dp"
                     android:layout_height="match_parent"
@@ -77,7 +77,7 @@
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toBottomOf="@id/homeButtonSectionMapExtendedBlock" />
 
-                <com.spendesk.grapes.component.content.summary.noneditable.SummaryBlockContentMapView
+                <com.spendesk.grapes.component.content.summary.block.SummaryBlockContentMapView
                     android:id="@+id/homeButtonSectionMapCompactBlock"
                     android:layout_width="0dp"
                     android:layout_height="match_parent"
@@ -135,7 +135,7 @@
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toTopOf="parent" />
 
-                <com.spendesk.grapes.component.content.summary.noneditable.SummaryBlockContentTextView
+                <com.spendesk.grapes.component.content.summary.block.SummaryBlockContentTextView
                     android:id="@+id/homeButtonSectionTextWithValueBlock"
                     android:layout_width="0dp"
                     android:layout_height="match_parent"
@@ -159,7 +159,7 @@
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toBottomOf="@id/homeButtonSectionTextWithValueBlock" />
 
-                <com.spendesk.grapes.component.content.summary.noneditable.SummaryBlockContentTextView
+                <com.spendesk.grapes.component.content.summary.block.SummaryBlockContentTextView
                     android:id="@+id/homeButtonSectionTextWithoutValueBlock"
                     android:layout_width="0dp"
                     android:layout_height="match_parent"
